Some more miscellaneous fixes
- Basic ref-counting to terminate GLFW
- Made Win-GenProjects.bat work from every directory
- Fix Build warnings to do with BufferElement Offset Type

+++ BakuEngine/src/Baku/Renderer/Buffer.h
@@ -33,7 +33,7 @@ namespace Baku
         ShaderDataType Type;
         std::string Name;
         uint32_t Size;
-        uint32_t Offset;
+        size_t Offset;
         bool Normalized;
 
         BufferElement() = default;
@@ -86,7 +86,7 @@ namespace Baku
     private:
         void CalculateOffsetAndStride()
         {
-            uint32_t offset = 0;
+            size_t offset = 0;
             m_Stride = 0;
             for (auto& element : m_Elements)
             {
